% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_k_function.R
\name{decomp_k}
\alias{decomp_k}
\title{Decomposition K}
\usage{
decomp_k(data, initial, final, time)
}
\arguments{
\item{data}{Data Frame, the data frame you want to alter or analyze}

\item{initial}{Numeric, the initial mass of the plots analyzed (grams)}

\item{final}{Character, the name of the column with final litter bag mass values after a certain amount of time}

\item{time}{Numeric, the number of days the litter bags decomposed}
}
\description{
This function is designed to handle litter bag data from fertilized and non-fertilized plots. It calculates k values from inital and final masses of litter bags.
}
