% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_mean_sd_function.R
\name{matrix_mean_sd}
\alias{matrix_mean_sd}
\title{Mean+SD Matrix}
\usage{
matrix_mean_sd(data, fert, gdLev, column)
}
\arguments{
\item{data}{Data Frame, the data frame you want to alter or analyze}

\item{fert}{Character, the name of the column that tells whether plots are fertilized or not (values in the column should be "F" and "C")}

\item{gdLev}{Character, the name of the column that tells the genetic diversity level of the plots}

\item{column}{Character, the name of the column with values you want to analyze in a matrix}
}
\description{
This function is designed to create a matrix of means and standard deviations of a column of the user's choosing. it compares the means between different fertilization and genetic diversity levels.
}
