% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_and_compare_function.R
\name{sep_and_compare}
\alias{sep_and_compare}
\title{Separate and Compare}
\usage{
sep_and_compare(data, fert, column1, column2)
}
\arguments{
\item{data}{Data Frame, the data frame you want to alter or analyze}

\item{fert}{Character, the name of the column that tells whether plots are fertilized or not (values in the column should be "F" and "C")}

\item{column1}{Character, the name of the column with numeric values that you want to analyze in an anova test}

\item{column2}{Character, the name of the column with factor values (2+ levels) you want to analyze in an anova test}
}
\description{
This function is designed to run anova tests to analyze if there is a difference in a variable between genetic diversity levels. This function returns the anova results for all plots, then separates them by fertilization level.
}
